# backend/openapi.json
{
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "https://galibot-backend.224mr3hws1t4.eu-de.codeengine.appdomain.cloud",
            "description": "Development server"
        }
    ],
    "info": {
        "title": "Reservation APIs",
        "version": "0.1.0"
    },
    "paths": {
        "/documents/{document_id}/availability": {
            "get": {
                "summary": "Check Availability",
                "description": "Check for document availability",
                "operationId": "check_availability_documents__document_id__availability_get",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/documents{document_id}/reserve": {
            "post": {
                "summary": "Reserve Document",
                "description": "Reserve a document",
                "operationId": "reserve_document_documents_document_id__reserve_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentReservationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/exhibition/{exhibit_id}": {
            "get": {
                "summary": "Get Available Seats",
                "description": "Check if an exhibition is sold out",
                "operationId": "get_available_seats_exhibition__exhibit_id__get",
                "parameters": [
                    {
                        "name": "exhibit_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Exhibit Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SeatReservationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/exhibition/{exhibit_id}/reserve": {
            "post": {
                "summary": "Reserve Seats",
                "description": "Reserve a place for an exhibition",
                "operationId": "reserve_seats_exhibition__exhibit_id__reserve_post",
                "parameters": [
                    {
                        "name": "exhibit_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Exhibit Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeatReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SeatReservationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "DocumentReservationRequest": {
                "properties": {
                    "document_id": {
                        "type": "string",
                        "title": "Document Id"
                    }
                },
                "type": "object",
                "required": [
                    "document_id"
                ],
                "title": "DocumentReservationRequest"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "SeatReservationRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "seats": {
                        "type": "integer",
                        "title": "Seats"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "seats"
                ],
                "title": "SeatReservationRequest"
            },
            "SeatReservationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "remaining_seats": {
                        "type": "integer",
                        "title": "Remaining Seats"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "remaining_seats"
                ],
                "title": "SeatReservationResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}